[refactor] #37 가계부 목록 페이지 상태관리를 mobX에서 useState로 변경
- 가계부 목록 페이지에서 mobX를 사용하기에는 단일페이지에서만 필요한 값들이 많고 컴포넌트의 깊이가 깊지않아서 useState가 낫다고 판단하여 변경
