+++ FE/src/components/AccountBook/AccountBookList/AccountBookList.scss
@@ -1,90 +1,37 @@
-.acbook__list {
-  margin-left: auto;
+.delete__button {
+  z-index: 1;
+  transform: translate(80px, -60px);
+  animation: cardShow 0.5s;
+}
 
-  div:nth-child(1) {
-    animation: cardShow 0.5s;
-  }
-  div:nth-child(2) {
-    animation: cardShow 0.6s;
-  }
-  div:nth-child(3) {
-    animation: cardShow 0.7s;
-  }
-  div:nth-child(4) {
-    animation: cardShow 0.8s;
-  }
-  div:nth-child(5) {
-    animation: cardShow 0.9s;
-  }
-  div:nth-child(6) {
-    animation: cardShow 1.1s;
-  }
+.acbook {
+  background-color: #f9e000;
+  height: 200px;
+  width: 340px;
+  margin: 20px 0px 20px 0px;
+  border: 1px solid transparent;
+  border-radius: 15px;
+  padding: 16px;
+  transition-duration: 0.3s;
+  transform: translateX(20%);
+  color: black;
+  cursor: pointer;
 
-  .delete__button {
-    z-index: 1;
-    transform: translate(80px, -60px);
-    animation: cardShow 0.5s;
-  }
-
-  .create__acbook {
-    background-color: #f9e000;
-    height: 200px;
-    width: 340px;
-    margin: 20px 0px 20px 0px;
-    border: 1px solid transparent;
-    border-radius: 15px;
-    padding: 16px;
+  &:hover {
+    transform: translateX(5%);
     transition-duration: 0.3s;
-    transform: translateX(20%);
-
-    > input {
-      height: 30px;
-      width: 260px;
-      text-decoration: none;
-      margin: 5px;
-    }
-
-    > textarea {
-      height: 70px;
-      width: 260px;
-      margin: 5px;
-    }
-
-    .create__cancle {
-      cursor: pointer;
-      width: 0px;
-    }
+    transition-timing-function: ease, ease;
   }
 
-  .acbook {
-    background-color: #f9e000;
-    height: 200px;
-    width: 340px;
-    margin: 20px 0px 20px 0px;
-    border: 1px solid transparent;
-    border-radius: 15px;
-    padding: 16px;
-    transition-duration: 0.3s;
-    transform: translateX(20%);
-    color: black;
-    cursor: pointer;
-
-    &:hover {
-      transform: translateX(5%);
-      transition-duration: 0.3s;
-      transition-timing-function: ease, ease;
+  @keyframes cardShow {
+    from {
+      opacity: 0;
+      transform: translateX(300px);
     }
 
-    @keyframes cardShow {
-      from {
-        opacity: 0;
-        transform: translateX(300px);
-      }
-
-      to {
-        opacity: 1;
-        transform: 0;
-      }
+    to {
+      opacity: 1;
+      transform: 0;
     }
   }
 }
