+++ FE/src/pages/AccountBook/AccountBookListPage.scss
@@ -2,4 +2,27 @@
   display: flex;
   flex-direction: row;
   padding: 90px 60px;
+
+  .acbook__list {
+    margin-left: auto;
+
+    div:nth-child(1) {
+      animation: cardShow 0.5s;
+    }
+    div:nth-child(2) {
+      animation: cardShow 0.6s;
+    }
+    div:nth-child(3) {
+      animation: cardShow 0.7s;
+    }
+    div:nth-child(4) {
+      animation: cardShow 0.8s;
+    }
+    div:nth-child(5) {
+      animation: cardShow 0.9s;
+    }
+    div:nth-child(6) {
+      animation: cardShow 1.1s;
+    }
+  }
 }
